requirements.txt
mcp
google-cloud-bigquery
python-dotenv
